+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+* Allow column values to be passed in the css file (e.g `ri-columns(16)`)
+* Default value can still be set in options
+
 # 0.1.0
 
 * Initial release.